chore: 🤖 move directories
